pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcp-pinecone-notes"
version = "0.1.0"
description = "MCP server for managing notes with pinecone vector database"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "Tim Buchinger", email = "" }
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "mcp[cli]>=1.2.1",
    "pinecone",
    "fastapi>=0.109.0",
    "pydantic>=2.6.0",
    "uvicorn>=0.27.0",
    "python-dotenv>=1.0.0",
    "nomic[local]>=3.4.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "black>=24.1.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
    "pytest-asyncio>=0.23.0",
]

[project.scripts]
mcp-pinecone-notes = "mcp_pinecone_notes:main"

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.ruff]
target-version = "py311"
line-length = 88
select = ["E", "F", "B", "I"]

[tool.mypy]
python_version = "3.11"
strict = true
